First attempt at Blockchain communication
A contract written in Solidity to interface and send / request information on the Ethereum network. 
The storage exists to house information that will be sent to the Ethereum network. 
The website (needs MetaMask extension to allow for the browser to communicate with any public / private Ethereum network) allows the user to input their information into a new node that would've been added into the dynamic blockchain.